requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.25
jinja2==3.1.2
python-multipart==0.0.6
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
itsdangerous>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0

